[DLinkSmartHome] extract HNAP from DLinkSmartPlug
extract HNAP implementation from DLinkSmartPlug into separate class.
